Add modules for generic model & chromosome code.
